requirements.txt
flask
python-dotenv
flask-cors 
transformers
hf_hub_ctranslate2


